added build data, 3 trails with 1 trip each

+++ app/src/main/java/com/example/hiker/TrailDatabase.java
@@ -15,12 +15,13 @@ import static android.content.ContentValues.TAG;
 
 public class TrailDatabase extends SQLiteOpenHelper {
 
-    private static final int VERSION = 1;
+    private static final int DATABASE_VERSION = 2;
     private static final String DATABASE_NAME = "trailDatabase.db";
 
     private static TrailDatabase mTraildb;
 
     public static TrailDatabase getInstance(Context context) {
+
         if (mTraildb == null) {
             mTraildb = new TrailDatabase(context);
         }
@@ -29,7 +30,7 @@ public class TrailDatabase extends SQLiteOpenHelper {
 
     //constructor
     private TrailDatabase(Context context){
-        super(context, DATABASE_NAME, null, VERSION);
+        super(context, DATABASE_NAME, null, DATABASE_VERSION);
     }
 
     private static final class TrailsTable {
@@ -65,7 +66,7 @@ public class TrailDatabase extends SQLiteOpenHelper {
 
         sqLiteDatabase.execSQL("create table " + TripsTable.TABLE + " (" +
                 TripsTable.COL_TR_ID + " integer, " +
-                TripsTable.COL_DATE + " int, " +
+                TripsTable.COL_DATE + " , " +
                 TripsTable.COL_DESCRIPTION + ", " +
                 "foreign key(" + TripsTable.COL_TR_ID + ") references " +
                 TrailsTable.TABLE + "(" + TrailsTable.COL_TRAIL_ID + ") on delete cascade, " +
@@ -77,15 +78,39 @@ public class TrailDatabase extends SQLiteOpenHelper {
                 "foreign key (" + ImagesTable.COL_TR_ID + ") references " +
                 TrailsTable.TABLE + "(" + TrailsTable.COL_TRAIL_ID + ") on delete cascade, " +
                 "primary key (" + ImagesTable.COL_TR_ID + ", " + ImagesTable.COL_FILENAME +  "))");
+        //BUILD_DATA: build data goes here. create the tables as objects and pass that object and the database instance to the add<element>OnBuild method.
+        //i was getting an error for recursively creating the database when trying to add elements from this class using the normal add<element> methods.
+        //this is because those methods need to instantiate a writable database within their scope, that writable database is already in this scope so calling them from here created that recursion.
         Log.d(TAG, "onCreate: db created");
+        Trail trail = new Trail("Art Loeb", 29, 8257, 011);
+        addTrailOnBuild(trail, sqLiteDatabase);
+        trail = new Trail ("Blood Mountain", 6, 1545, 001);
+        addTrailOnBuild(trail, sqLiteDatabase);
+        trail = new Trail ("Bear Creek", 6, 1108, 011);
+        addTrailOnBuild(trail, sqLiteDatabase);
+        Trip trip = new Trip (1);
+        trip.setDate("07/11/2011");
+        trip.setDescription("Went with Corn and Shelby. made a wrong turn which added 10 miles to the journey.");
+        addTripOnBuild(trip, sqLiteDatabase);
+        trip = new Trip(2);
+        trip.setDate("8/12/2018");
+        trip.setDescription("Rained the whole time but was still beautiful");
+        addTripOnBuild(trip, sqLiteDatabase);
+        trip = new Trip(3);
+        trip.setDate("04/10/2014");
+        trip.setDescription("Went with brown and kumu. Almost ate jeffrey the turtle but he ran away");
+        addTripOnBuild(trip, sqLiteDatabase);
     }
 
     @Override
-    public void onUpgrade(SQLiteDatabase sqLiteDatabase, int i, int i1) {
-        sqLiteDatabase.execSQL("drop table if exists " + TrailsTable.TABLE);
-        sqLiteDatabase.execSQL("drop table if exists " + TripsTable.TABLE);
-        sqLiteDatabase.execSQL("drop table if exists " + ImagesTable.TABLE);
-        onCreate(sqLiteDatabase);
+    public void onUpgrade(SQLiteDatabase sqLiteDatabase, int oldVersion, int newVersion) {
+        if (newVersion>oldVersion) {
+            sqLiteDatabase.execSQL("drop table if exists " + TrailsTable.TABLE);
+            sqLiteDatabase.execSQL("drop table if exists " + TripsTable.TABLE);
+            sqLiteDatabase.execSQL("drop table if exists " + ImagesTable.TABLE);
+            onCreate(sqLiteDatabase);
+            Log.d(TAG, "onUgrade: upgraded to version " + DATABASE_VERSION);
+        }
     }
 
     @Override
@@ -93,6 +118,22 @@ public class TrailDatabase extends SQLiteOpenHelper {
         super.onOpen(sqLiteDatabase);
         if (!sqLiteDatabase.isReadOnly()) {
             sqLiteDatabase.setForeignKeyConstraintsEnabled(true);
+        }
+        Log.d(TAG,"opened on version " + DATABASE_VERSION);
+    }
+
+    private void addTrailOnBuild(Trail trail, SQLiteDatabase db) {
+        ContentValues values = new ContentValues();
+        values.put(TrailsTable.COL_TRAIL_NAME, trail.getTrail_name());
+        values.put(TrailsTable.COL_ELEVATION, trail.getElevation());
+        values.put(TrailsTable.COL_FEATURES, trail.getFeatures());
+        values.put(TrailsTable.COL_TRAIL_DISTANCE, trail.getTrail_distance());
+        long trailId = db.insert(TrailsTable.TABLE, null, values);
+        if (trailId == -1) {
+            Log.println(Log.INFO, "error:", trail.getTrail_name());
+        } else {
+            trail.setTrail_id(trailId);
+            Log.d(TAG, "addTrail: id = "+ trailId);
         }
     }
 
@@ -108,8 +149,24 @@ public class TrailDatabase extends SQLiteOpenHelper {
             Log.println(Log.INFO, "error:", trail.getTrail_name());
         } else {
             trail.setTrail_id(trailId);
-            Log.d(TAG, "addTrail: worked");
+            Log.d(TAG, "addTrail: id = "+ trailId);
         }
+    }
+
+    private void addTripOnBuild (Trip trip, SQLiteDatabase db) {
+        ContentValues values = new ContentValues();
+        values.put(TripsTable.COL_TR_ID, trip.getTr_id());
+        values.put(TripsTable.COL_DATE, trip.getDate());
+        values.put(TripsTable.COL_DESCRIPTION, trip.getDescription());
+        db.insert(TripsTable.TABLE, null, values);
+    }
+    public void addTrip (Trip trip) {
+        SQLiteDatabase db = getWritableDatabase();
+        ContentValues values = new ContentValues();
+        values.put(TripsTable.COL_TR_ID, trip.getTr_id());
+        values.put(TripsTable.COL_DATE, trip.getDate());
+        values.put(TripsTable.COL_DESCRIPTION, trip.getDescription());
+        db.insert(TripsTable.TABLE, null, values);
     }
 
     public List<String> getTrails(int distanceUpper, int distanceLower){
